Cambios en cociente (inverso)

+++ Actividad_02_Entornos_de_Desarrollo/src/Cociente.java
@@ -67,8 +67,8 @@ public class Cociente {
 	 * 
 	 */
 	
-	public int cocienteInverso (int a) {
-		return 1 / a;
+	public int cocienteInverso (double a) {
+		return (int) (1 / a);
 	}
 	
 	/**
